app/bot/requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
openai==1.45.0
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
python-dotenv==1.0.1
pdfplumber==0.11.4